ReadFrustrationData
The Controller sets the blackboard with the path to Devin's frustration file using vars for key name and value.
BTDecorator_ReadFrustrationData is responsible for checking if the file exists.
Returns true if the file exists.
Updated BT for Devin's method to check BTDecorator_ReadFrustrationData and set frustration C (for now) if true.
Need to check the actual contents of the data file next.